refactor: split codebase into multiple modules
