+++ src/defaults.rs
@@ -1,0 +1,8 @@
+pub const IDLE_TIME: u64 = 5;
+pub const ACTIVE_TIME: u64 = 1;
+
+pub const DISCORD_ID: u64 = 677226551607033903;
+pub const DEFAULT_HOST: &str = "localhost:6600";
+
+pub const DETAILS_FORMAT: &str = "$title";
+pub const STATE_FORMAT: &str = "$artist / $album";
